Add: Reducer function to update state
The Reducer function is responsible to updating the state based on the current action. It takes alwas to arguments, the "state" and the "action". This has to be a pure function to increase predictability. So the function doesn't modify the current, instead with the concat() method it creates a new one, in order to remain pure.

When we first will call the function, the state is gonna be undefined. So we should use "ES6 default paramters" (state=[]), so if state is undefined, it set that to an empty array, which can perform the concat method to it.