Feat: Add dispatch function to the state.
Dispatch function is for to update the state inside of the store.
In order to do it asks for the value of the reducer function and sets it as a new state.
The dispatch function also calls all the registered listeners, because the state has potentially changed
As todos was an app code, while store is our library code, we passed the former as a paramter order to handle later access issues