refactor: Switch to a check middleWare Redux function insted ouf our checkAndDispatch().
The middleware is a higher order function which is applied to the store and invoked right between the dispatch and the reducer. With that we can invoke the regular dispatch function and be able to implement as many middleware function as we just want without rebuilding our function-invoke system every time.